Rename "ConfirmedCasesNumberCard" to "PatientsTransitionCard"
