+++ components/cards/PatientsTransitionCard.vue
